// src/flowchart/helpers/flowchart-steps.enum.ts
export enum FlowchartStepsEnum {
  STEP_INITIAL = 'STEP_INITIAL',
  STEP_SCRIPT = 'STEP_SCRIPT',
  STEP_REQUEST = 'STEP_REQUEST',
  STEP_NOTIFICATION = 'STEP_NOTIFICATION',
  STEP_FLOW_REQUEST = 'STEP_FLOW_REQUEST',
  STEP_LOOP_START = 'STEP_LOOP_START',
  STEP_LOOP_END = 'STEP_LOOP_END',
  STEP_CONDITIONAL = 'STEP_CONDITIONAL',
  STEP_FINALIZATION = 'STEP_FINALIZATION',
  STEP_COLLECTION = 'STEP_COLLECTION',
  STEP_RESULT = 'STEP_RESULT',
}